Load and unload music files as needed
When VVVVVV is run on devices (such as the Nintendo Wii) that are memory
costrained, it's not possible to load all the music files at startup.

Therefore, change the MusicTrack contructor to only save the RWops
structure, and load the music data only when Play() gets called. Also,
when a song start playing we do unload the previous one.